requirements.txt
pandas
faiss-cpu
sentence-transformers
groq
streamlit
ragas
langchain-groq
langchain-community
datasets
numpy
torch
transformers